requirements.txt
keyboard==0.13.5 
pyaes==1.6.1
pykeyboard==0.1.5
Pyrogram==2.0.106
pyserial==3.5
PySocks==1.7.1
TgCrypto==1.2.5
picamera2==0.3.27
